Group manager added

+++ src/com/soc/components/Health.java
@@ -3,8 +3,8 @@ package com.soc.components;
 import com.artemis.Component;
 
 public class Health extends Component {
-	double health;
-	public Health(double health){
+	int health;
+	public Health(int health){
 		this.health=health;
 	}
 }
